pyproject.toml
[project]
name = "hashicorp-code-along"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.3.2",
    "fastapi[standard]>=0.116.1",
    "hvac>=2.3.0",
    "python-dotenv>=1.1.1",
]
